# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Riise Map API
  version: 1.0.0
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /learners/{learnerId}/projects:
    get:
      operationId: getLearnerProjects
      tags: [learners]
      summary: Get projects for a specific learner
      parameters:
        - name: learnerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A list of learner projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LearnerProject"
...
components:
  schemas:
...
    LearnerProject:
      type: object
      properties:
        id:
          type: integer
        learnerId:
          type: integer
        title:
          type: string
        completion:
          type: integer
        status:
          type: string
      required:
        - id
        - learnerId
        - title
        - completion
        - status